% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BurnArea-package.R
\docType{package}
\name{BurnArea-package}
\alias{BurnArea-package}
\title{BurnArea: Analysis of Burned Areas Using Optical Satellite Imagery}
\description{
Tools for analyzing burned areas using optical satellite imagery. This package provides functions for NBR calculation, burn severity classification, and validation of results.
}
\author{
\strong{Maintainer}: Fajar Raihan \email{fajar.raihan.data@gmail.com}

}
\keyword{internal}
