% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calculate_burn_stats}
\alias{calculate_burn_stats}
\title{Calculate Burn Area Statistics}
\usage{
calculate_burn_stats(severity_raster, training_data, pixel_size = 0.09)
}
\arguments{
\item{severity_raster}{Raster of classified burn severity}

\item{training_data}{Training data for statistics calculation}

\item{pixel_size}{Pixel size in hectares (default = 0.09 for 30m Landsat resolution)}
}
\value{
List containing burn area statistics
}
\description{
Calculate Burn Area Statistics
}
