% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validasi_D_value.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate Distance Index for NBR}
\usage{
calculate_distance(training_data)
}
\arguments{
\item{training_data}{Training data for threshold calculation}
}
\value{
Distance index value
}
\description{
Calculate Distance Index for NBR
}
