% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koreksi_radiometric.R
\name{calculate_earth_sun_distance}
\alias{calculate_earth_sun_distance}
\title{Calculate Earth-Sun Distance}
\usage{
calculate_earth_sun_distance(doy)
}
\arguments{
\item{doy}{Day of year (1-366)}
}
\value{
Earth-Sun distance in astronomical units
}
\description{
Calculate Earth-Sun Distance
}
