% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbr_calculation.R
\name{calculate_nbr}
\alias{calculate_nbr}
\title{Calculate Normalized Burn Ratio (NBR)}
\usage{
calculate_nbr(preprocessed_stack)
}
\arguments{
\item{preprocessed_stack}{Preprocessed raster stack}
}
\value{
NBR raster
}
\description{
Calculate Normalized Burn Ratio (NBR)
}
