% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_severity}
\alias{classify_severity}
\title{Classify Burn Severity Levels Based on Burned Area}
\usage{
classify_severity(nbr_post, dnbr, training_data)
}
\arguments{
\item{nbr_post}{Post-fire NBR raster}

\item{dnbr}{Delta NBR raster}

\item{training_data}{Training data for threshold calculation}
}
\value{
Classified raster of burn severity
}
\description{
Classify Burn Severity Levels Based on Burned Area
}
