% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_burns}
\alias{export_burns}
\title{Export Burn Area Results}
\usage{
export_burns(burn_raster, filename, format = "GTiff")
}
\arguments{
\item{burn_raster}{Burn area raster}

\item{filename}{Output filename}

\item{format}{Output format ("GTiff", "ASCII", "ESRI Shapefile")}
}
\description{
Export Burn Area Results
}
