% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_data.R
\name{generate_training_data}
\alias{generate_training_data}
\title{Generate Training Data from Polygon Shapefile}
\usage{
generate_training_data(nbr_pre, nbr_post, dnbr, training_poly)
}
\arguments{
\item{nbr_pre}{Raster NBR pre-fire}

\item{nbr_post}{Raster NBR post-fire}

\item{dnbr}{Raster dNBR}

\item{training_poly}{SpatialPolygonsDataFrame berisi area training}
}
\value{
List berisi training data pre, post fire dan dNBR
}
\description{
Generate Training Data from Polygon Shapefile
}
