% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_burned_areas.R
\name{identify_burned_areas}
\alias{identify_burned_areas}
\title{Identify Burned Areas Using NBR Thresholds}
\usage{
identify_burned_areas(nbr_post, dnbr, training_data)
}
\arguments{
\item{nbr_post}{Post-fire NBR raster}

\item{dnbr}{Delta NBR raster}

\item{training_data}{Training data for threshold calculation}
}
\value{
Raster of identified burned areas
}
\description{
Identify Burned Areas Using NBR Thresholds
}
