% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi_uji_validasi_icsi.R
\name{plot_burn_accuracy}
\alias{plot_burn_accuracy}
\title{Plot Validation}
\usage{
plot_burn_accuracy(predicted_raster, reference_shp)
}
\arguments{
\item{predicted_raster}{Raster of predicted burn area (binary: 1 for burned, 0 for unburned)}

\item{reference_shp}{Shapefile of reference burn area (polygon)}
}
\value{
List containing omission error, commission error, and ICSI value
}
\description{
Plot Validation
}
