% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisasi.R
\name{plot_burn_severity}
\alias{plot_burn_severity}
\title{Plot Burn Severity Classification}
\usage{
plot_burn_severity(
  severity_raster,
  title = "Burn Severity Classification",
  save_plot = FALSE,
  output_path = NULL
)
}
\arguments{
\item{severity_raster}{Raster hasil klasifikasi dari classify_severity}

\item{title}{Judul plot (optional)}

\item{save_plot}{Boolean untuk menyimpan plot (default: FALSE)}

\item{output_path}{Path untuk menyimpan plot (optional)}
}
\value{
Plot tingkat keparahan kebakaran
}
\description{
Plot Burn Severity Classification
}
