% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_nbr}
\alias{preprocess_nbr}
\title{Preprocessing Data for NBR Analysis}
\usage{
preprocess_nbr(nir_band, swir_band, mask = NULL)
}
\arguments{
\item{nir_band}{Near-infrared band raster}

\item{swir_band}{Short-wave infrared band raster}

\item{mask}{Optional mask raster}
}
\value{
Preprocessed raster stack
}
\description{
Preprocessing Data for NBR Analysis
}
