% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_to_raster_reference.R
\name{shp_to_raster_reference}
\alias{shp_to_raster_reference}
\title{Convert shp reference to reference raster}
\usage{
shp_to_raster_reference(reference_shp, predicted_raster)
}
\arguments{
\item{reference_shp}{Shapefile of reference burn area (polygon)}

\item{predicted_raster}{Raster of predicted burn area (binary: 1 for burned, 0 for unburned)}
}
\value{
List containing omission error, commission error, and ICSI value
}
\description{
Convert shp reference to reference raster
}
