% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referensi_terbakar.R
\name{upload_reference_shp}
\alias{upload_reference_shp}
\title{Upload Reference Shapefile Data}
\usage{
upload_reference_shp(shp_path, crs_proj = NULL)
}
\arguments{
\item{shp_path}{Character string path to shapefile}

\item{crs_proj}{Projection string (optional)}
}
\value{
SpatialPolygonsDataFrame object
}
\description{
Upload Reference Shapefile Data
}
