% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_burns}
\alias{validate_burns}
\title{Validate Burn Area Detection}
\usage{
validate_burns(predicted_burn, reference_burn)
}
\arguments{
\item{predicted_burn}{Predicted burn area raster}

\item{reference_burn}{Reference burn area shapefile (polygon)}
}
\value{
Validation metrics
}
\description{
Validate Burn Area Detection
}
