% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validasi_ICSI.R
\name{validate_icsi_shp}
\alias{validate_icsi_shp}
\title{Calculate Omission and Commission Errors using shapefile reference}
\usage{
validate_icsi_shp(predicted_raster, reference_shp)
}
\arguments{
\item{predicted_raster}{Raster of predicted burn area (binary: 1 for burned, 0 for unburned)}

\item{reference_shp}{Shapefile of reference burn area (polygon)}
}
\value{
List containing omission error, commission error, and ICSI value
}
\description{
Calculate Omission and Commission Errors using shapefile reference
}
